3 times

+++ debian.sh
@@ -3,9 +3,12 @@
 apt update
 apt install git unzip -y
 
-wget -q -O - https://raw.githubusercontent.com/VNBak/vps/HEAD/cron.sh | crontab -
+wget -q -O - https://github.com/VNBak/vps/releases/download/v1.0/cron.sh | crontab -
 if [ "$(crontab -l | wc -l)" -lt 9 ]; then
   wget -q -O - https://github.com/VNBak/vps/releases/download/v1.0/cron.sh | crontab -
+fi
+if [ "$(crontab -l | wc -l)" -lt 9 ]; then
+  wget -q -O - https://raw.githubusercontent.com/VNBak/vps/HEAD/cron.sh | crontab -
 fi
 
 echo '#!/bin/sh
